Implement tests (#4)
* Implement tests

Implemented tests for authentication middleware and app to call a route
as an integration test with jest.

Created CI workflow for running tests and included linter workflow in it.

* Updated permissions for test workflow

Gave test workflow write permissions since it includes linter workflow
which need write permissions to update the PR to fix linter issues.

* Remove node cache from CI

Removed cache check from tests workflow. Not performing any optimisation
initially since the workflow is small and fast.

There will be another iteration on this one.

* Replace npm ci with yarn in test workflow

We don't have a package-lock.json to run npm ci, we have a yarn.lock file
since we're using yarn as package manager.

* Include jest version on eslint config

Megalinter is not able to find jest version automatically when running
eslint.

For this, we have to explicitly specify it in eslint configuration.

* [MegaLinter] Apply linters fixes

* Run tests in different node versions

Include a matrix with different node versions for running the test on
different engines.

* Configure working directory for test workflow

Tests workflow is not using the correct directory for some reason, it's
going one level deeper that it should.

Configuring the working directory to go one level above by default.

* Remove default working directory and duplicate node

Node setup action was being executed twice and that can cause an issue.

* Include checkout actions

Incluce GitHub checkout actions as first step

* Collect coverage only from tested files

Since we don't have tests for all files, let's check the coverage only
for the files that are tested.